Simplify Iceberg metadata loading structure
