+++ server/sessionmeta/sessionmeta.go
@@ -226,9 +226,51 @@ func buildSessionPgDatabaseViewSQL(database string) string {
 func buildSessionInformationSchemaColumnsViewSQL() string {
 	return `
 		CREATE OR REPLACE VIEW main.information_schema_columns_compat AS
+		WITH all_columns AS (
+			SELECT
+				CASE WHEN c.table_catalog IN ('ducklake', 'memory') THEN current_database() ELSE c.table_catalog END AS table_catalog,
+				CASE WHEN c.table_schema = 'main' THEN 'public' ELSE c.table_schema END AS table_schema,
+				c.table_name,
+				c.column_name,
+				c.ordinal_position,
+				c.column_default,
+				c.is_nullable,
+				c.data_type,
+				COALESCE(m.character_maximum_length, c.character_maximum_length) AS character_maximum_length,
+				c.character_octet_length,
+				COALESCE(m.numeric_precision, c.numeric_precision) AS numeric_precision,
+				COALESCE(m.numeric_scale, c.numeric_scale) AS numeric_scale,
+				c.datetime_precision
+			FROM information_schema.columns c
+			LEFT JOIN main.__duckgres_column_metadata m
+				ON c.table_schema = m.table_schema
+				AND c.table_name = m.table_name
+				AND c.column_name = m.column_name
+			WHERE NOT (
+				c.table_catalog = 'iceberg'
+				AND c.column_name = '__'
+				AND UPPER(c.data_type) = 'UNKNOWN'
+			)
+			UNION ALL
+			SELECT
+				'iceberg' AS table_catalog,
+				table_schema,
+				table_name,
+				column_name,
+				ordinal_position,
+				NULL AS column_default,
+				is_nullable,
+				data_type,
+				character_maximum_length,
+				character_octet_length,
+				numeric_precision,
+				numeric_scale,
+				datetime_precision
+			FROM main.__duckgres_iceberg_column_metadata
+		)
 		SELECT
-			CASE WHEN c.table_catalog IN ('ducklake', 'memory') THEN current_database() ELSE c.table_catalog END AS table_catalog,
-			CASE WHEN c.table_schema = 'main' THEN 'public' ELSE c.table_schema END AS table_schema,
+			c.table_catalog,
+			c.table_schema,
 			c.table_name,
 			c.column_name,
 			c.ordinal_position,
@@ -245,14 +287,15 @@ func buildSessionInformationSchemaColumnsViewSQL() string {
 				WHEN UPPER(c.data_type) = 'VARCHAR' OR UPPER(c.data_type) LIKE 'VARCHAR(%' THEN 'text'
 				WHEN UPPER(c.data_type) = 'TEXT' THEN 'text'
 				WHEN UPPER(c.data_type) LIKE 'TEXT(%' THEN 'character'
-				WHEN UPPER(c.data_type) = 'BOOLEAN' THEN 'boolean'
+				WHEN UPPER(c.data_type) = 'STRING' THEN 'text'
+				WHEN UPPER(c.data_type) = 'BOOLEAN' OR UPPER(c.data_type) = 'BOOL' THEN 'boolean'
 				WHEN UPPER(c.data_type) = 'TINYINT' THEN 'smallint'
 				WHEN UPPER(c.data_type) = 'SMALLINT' THEN 'smallint'
-				WHEN UPPER(c.data_type) = 'INTEGER' THEN 'integer'
-				WHEN UPPER(c.data_type) = 'BIGINT' THEN 'bigint'
+				WHEN UPPER(c.data_type) = 'INTEGER' OR UPPER(c.data_type) = 'INT' THEN 'integer'
+				WHEN UPPER(c.data_type) = 'BIGINT' OR UPPER(c.data_type) = 'LONG' THEN 'bigint'
 				WHEN UPPER(c.data_type) = 'HUGEINT' THEN 'numeric'
 				WHEN UPPER(c.data_type) = 'REAL' OR UPPER(c.data_type) = 'FLOAT4' THEN 'real'
-				WHEN UPPER(c.data_type) = 'DOUBLE' OR UPPER(c.data_type) = 'FLOAT8' THEN 'double precision'
+				WHEN UPPER(c.data_type) = 'DOUBLE' OR UPPER(c.data_type) = 'FLOAT8' OR UPPER(c.data_type) = 'DOUBLE PRECISION' THEN 'double precision'
 				WHEN UPPER(c.data_type) LIKE 'DECIMAL%' THEN 'numeric'
 				WHEN UPPER(c.data_type) LIKE 'NUMERIC%' THEN 'numeric'
 				WHEN UPPER(c.data_type) = 'DATE' THEN 'date'
@@ -261,15 +304,16 @@ func buildSessionInformationSchemaColumnsViewSQL() string {
 				WHEN UPPER(c.data_type) = 'TIMESTAMPTZ' OR UPPER(c.data_type) = 'TIMESTAMP WITH TIME ZONE' THEN 'timestamp with time zone'
 				WHEN UPPER(c.data_type) = 'INTERVAL' THEN 'interval'
 				WHEN UPPER(c.data_type) = 'UUID' THEN 'uuid'
-				WHEN UPPER(c.data_type) = 'BLOB' OR UPPER(c.data_type) = 'BYTEA' THEN 'bytea'
+				WHEN UPPER(c.data_type) = 'BLOB' OR UPPER(c.data_type) = 'BYTEA' OR UPPER(c.data_type) = 'BINARY' OR UPPER(c.data_type) = 'FIXED' THEN 'bytea'
 				WHEN UPPER(c.data_type) = 'JSON' THEN 'json'
-				WHEN UPPER(c.data_type) LIKE '%[]' THEN 'ARRAY'
+				WHEN UPPER(c.data_type) LIKE '%[]' OR UPPER(c.data_type) LIKE 'LIST%' THEN 'ARRAY'
+				WHEN UPPER(c.data_type) LIKE 'STRUCT%' OR UPPER(c.data_type) LIKE 'MAP%' OR c.data_type LIKE '{%' THEN 'json'
 				ELSE LOWER(c.data_type)
 			END AS data_type,
-			COALESCE(m.character_maximum_length, c.character_maximum_length) AS character_maximum_length,
+			c.character_maximum_length,
 			c.character_octet_length,
-			COALESCE(m.numeric_precision, c.numeric_precision) AS numeric_precision,
-			COALESCE(m.numeric_scale, c.numeric_scale) AS numeric_scale,
+			c.numeric_precision,
+			c.numeric_scale,
 			c.datetime_precision,
 			NULL AS interval_type,
 			NULL AS interval_precision,
@@ -301,108 +345,7 @@ func buildSessionInformationSchemaColumnsViewSQL() string {
 			'NEVER' AS is_generated,
 			NULL AS generation_expression,
 			'YES' AS is_updatable
-		FROM information_schema.columns c
-		LEFT JOIN main.__duckgres_column_metadata m
-			ON c.table_schema = m.table_schema
-			AND c.table_name = m.table_name
-			AND c.column_name = m.column_name
-		WHERE NOT (
-			c.table_catalog = 'iceberg'
-			AND c.column_name = '__'
-			AND UPPER(c.data_type) = 'UNKNOWN'
-		)
-		UNION ALL
-		SELECT
-			table_catalog,
-			table_schema,
-			table_name,
-			column_name,
-			ordinal_position,
-			column_default,
-			is_nullable,
-			data_type,
-			character_maximum_length,
-			character_octet_length,
-			numeric_precision,
-			numeric_scale,
-			datetime_precision,
-			interval_type,
-			interval_precision,
-			character_set_catalog,
-			character_set_schema,
-			character_set_name,
-			collation_catalog,
-			collation_schema,
-			collation_name,
-			domain_catalog,
-			domain_schema,
-			domain_name,
-			udt_catalog,
-			udt_schema,
-			udt_name,
-			scope_catalog,
-			scope_schema,
-			scope_name,
-			maximum_cardinality,
-			dtd_identifier,
-			is_self_referencing,
-			is_identity,
-			identity_generation,
-			identity_start,
-			identity_increment,
-			identity_maximum,
-			identity_minimum,
-			identity_cycle,
-			is_generated,
-			generation_expression,
-			is_updatable
-		FROM (
-			SELECT
-				'iceberg' AS table_catalog,
-				table_schema,
-				table_name,
-				column_name,
-				ordinal_position,
-				NULL AS column_default,
-				is_nullable,
-				data_type,
-				character_maximum_length,
-				character_octet_length,
-				numeric_precision,
-				numeric_scale,
-				datetime_precision,
-				NULL AS interval_type,
-				NULL AS interval_precision,
-				NULL AS character_set_catalog,
-				NULL AS character_set_schema,
-				NULL AS character_set_name,
-				NULL AS collation_catalog,
-				NULL AS collation_schema,
-				NULL AS collation_name,
-				NULL AS domain_catalog,
-				NULL AS domain_schema,
-				NULL AS domain_name,
-				NULL AS udt_catalog,
-				NULL AS udt_schema,
-				NULL AS udt_name,
-				NULL AS scope_catalog,
-				NULL AS scope_schema,
-				NULL AS scope_name,
-				NULL AS maximum_cardinality,
-				NULL AS dtd_identifier,
-				'NO' AS is_self_referencing,
-				'NO' AS is_identity,
-				NULL AS identity_generation,
-				NULL AS identity_start,
-				NULL AS identity_increment,
-				NULL AS identity_maximum,
-				NULL AS identity_minimum,
-				NULL AS identity_cycle,
-				'NEVER' AS is_generated,
-				NULL AS generation_expression,
-				'YES' AS is_updatable
-			FROM main.__duckgres_iceberg_column_metadata
-		)
+		FROM all_columns c
 	`
 }
